--- Assets/GameLogic/AnimationBomb.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fd4e6d3a21f032a469ff3aa892e66941
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
